fix: support desktop files
Support for desktop files is added by using a listview to get the base (extensionless) name of the file and then going through the directory of all extensionless files and finding one with a matching name.
This will not work well when there are files with the same name but a different extension. That is a problem for later.